refactor: move `ParamType` and `Enum` encoding functions to fuels-types
